+++ zview/ztext.h
@@ -1,4 +1,4 @@
-extern void size_to_text( char *txt, float size);
+extern void size_to_text( char *txt, uint32 size);
 extern void str2lower(char *str);
 extern void str2upper(char *str);
 extern char *zstrncpy( char *dst, const char *src, size_t n);
